menu left + top v1
