+++ src/AppBundle/Controller/MenuController.php
@@ -24,29 +24,28 @@ class MenuController extends Controller
         // Get menus
         $menus = $this->container->getParameter("leftMenu");
 
-        $authManager = $this->get('nouestil.user');
-
+        $userManager = $this->get('nouestil.user');
         foreach ($menus as $key => $menu) {
-            if (!$authManager->isGranted('ROLE_ADMIN',$this->getUser()) && $menu['scope'] == "admin") {
+            if (!$userManager->isAdmin($this->getUser()) && $menu['scope'] == "admin") {
                 unset($menus[$key]);
             }
 
-            if (!$authManager->isGranted('ROLE_USER',$this->getUser()) && $menu['scope'] == "user") {
+            if (!$userManager->isUser($this->getUser()) && $menu['scope'] == "user") {
                 unset($menus[$key]);
             }
-            if(isset($menu['items'])){
-                foreach($menu["items"] as $subKey => $subMenu){
-                    if(isset($subMenu['scope'])){
-                        if (!$authManager->isGranted('ROLE_ADMIN',$this->getUser()) && $this->getUser()->getGroup()->getCode()!="pwp" && $subMenu['scope'] == "pwp") {
-                            unset($menus[$key]['items'][$subKey]);
-                        }
-                        if (!$this->getUser()->getTimeTracking() && $subMenu['scope'] == "time-tracking") {
-                            unset($menus[$key]['items'][$subKey]);
-                        }
-                    }
-
-                }
-            }
+//            if(isset($menu['items'])){
+//                foreach($menu["items"] as $subKey => $subMenu){
+//                    if(isset($subMenu['scope'])){
+//                        if (!$userManager->isGranted('ROLE_ADMIN',$this->getUser()) && $this->getUser()->getGroup()->getCode()!="pwp" && $subMenu['scope'] == "pwp") {
+//                            unset($menus[$key]['items'][$subKey]);
+//                        }
+//                        if (!$this->getUser()->getTimeTracking() && $subMenu['scope'] == "time-tracking") {
+//                            unset($menus[$key]['items'][$subKey]);
+//                        }
+//                    }
+//
+//                }
+//            }
 
         }
 
@@ -55,7 +54,7 @@ class MenuController extends Controller
         $parentRequest = $this->container->get('request_stack')->getParentRequest();
 
         // Return
-        return $this->render('PortailMkgBundle:Partials:sidebar.menu.html.twig', ['menu' => $menus, 'parentRequest' => $parentRequest]);
+        return $this->render('AppBundle:Partials:left.menu.html.twig', ['menu' => $menus, 'parentRequest' => $parentRequest]);
 
     }
 
@@ -70,7 +69,7 @@ class MenuController extends Controller
         $params = $this->container->getParameter("topbar");
 
         // Return
-        return $this->render('PortailMkgBundle:Partials:topbar.menu.html.twig', ['config' => $params]);
+        return $this->render('AppBundle:Partials:top.menu.html.twig', ['config' => $params]);
 
     }
