+++ src/AppBundle/Manager/UserManager.php
@@ -5,7 +5,6 @@ namespace AppBundle\Manager;
 use AppBundle\Entity\Contact;
 use AppBundle\Entity\User;
 use Doctrine\ORM\EntityManager;
-use Symfony\Component\HttpFoundation\Session\Session;
 
 /**
  * Class UserManager
@@ -35,26 +34,18 @@ class UserManager
     }
 
     /**
-     * @return array
+     * @return bool
      */
-    public function getRoles()
+    public function isAdmin($user)
     {
-        $roles = array();
-        if (is_array($this->roles) && sizeof($this->roles) > 0) {
-            foreach ($this->roles as $role) {
-                $roles[] = $role->name;
-            }
-        }
-
-        return $roles;
+        return (in_array('ROLE_ADMIN', $user->getRoles())) ? true : false;
     }
-
     /**
      * @return bool
      */
-    public function isAdmin()
+    public function isUser($user)
     {
-        return (in_array('admin', $this->getRoles())) ? true : false;
+        return (in_array('ROLE_USER', $user->getRoles())) ? true : false;
     }
 
 
@@ -217,38 +208,6 @@ class UserManager
         $this->em->flush();
     }
 
-
-    /**
-     * @return array
-     */
-    public function getContacts()
-    {
-        $contacts = array();
-        if (is_array($this->contacts) && sizeof($this->contacts) > 0) {
-            foreach ($this->contacts as $contact) {
-                $users[] = $contact->name;
-            }
-        }
-
-        return $contacts;
-    }
-
-    /**
-     * @return array
-     */
-    public function getContactById($id)
-    {
-        $contact = $this->em
-            ->getRepository('AppBundle:Contact')
-            ->findOneById($id);
-
-        if (!$contact instanceof Contact) {
-            throw $this->createNotFoundException(
-                'N\'est pas un contact'
-            );
-        }
-        return $contact;
-    }
 
     public function unlinkContact($userId, $contactId){
         $this->getUser($userId)->removeContact($contactId);
