+++ src/AppBundle/Resources/views/Profile/show.html.twig
@@ -1,7 +1,6 @@
 {% extends 'base.html.twig' %}
 
 {% block body %}
-    {% endfor %}
     {% block fos_user_content %}
         {% include "@FOSUser/Profile/show_content.html.twig" %}
     {% endblock fos_user_content %}
